@@ -1,7 +1,22 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.8
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
+
+let debugOtherSwiftFlags = [
+    "-Xfrontend", "-warn-long-expression-type-checking=500",
+    "-Xfrontend", "-warn-long-function-bodies=500",
+    "-strict-concurrency=complete",
+    "-enable-actor-data-race-checks",
+]
+
+let debugSwiftSettings: [PackageDescription.SwiftSetting] = [ // Swift5.8
+    .unsafeFlags(debugOtherSwiftFlags, .when(configuration: .debug)),
+    .enableUpcomingFeature("ConciseMagicFile", .when(configuration: .debug)), // SE-0274
+    .enableUpcomingFeature("ForwardTrailingClosures", .when(configuration: .debug)), // SE-0286
+    .enableUpcomingFeature("ExistentialAny", .when(configuration: .debug)), // SE-0335
+    .enableUpcomingFeature("BaseSlashRegexLiterals", .when(configuration: .debug)), // SE-0354
+]
 
 let package = Package(
     name: "Package",
